Lần này test ver3

+++ src/red/jdbc/daos/GodGK.java
@@ -1667,13 +1667,13 @@ public class GodGK {
                     pet.nPoint.tiemNang = Long.parseLong(String.valueOf(dataArray.get(2)));
                     pet.nPoint.stamina = Short.parseShort(String.valueOf(dataArray.get(3)));
                     pet.nPoint.maxStamina = Short.parseShort(String.valueOf(dataArray.get(4)));
-                    pet.nPoint.hpg = Integer.parseInt(String.valueOf(dataArray.get(5)));
-                    pet.nPoint.mpg = Integer.parseInt(String.valueOf(dataArray.get(6)));
-                    pet.nPoint.dameg = Integer.parseInt(String.valueOf(dataArray.get(7)));
+                    pet.nPoint.hpg = Long.parseLong(String.valueOf(dataArray.get(5)));
+                    pet.nPoint.mpg = Long.parseLong(String.valueOf(dataArray.get(6)));
+                    pet.nPoint.dameg = Long.parseLong(String.valueOf(dataArray.get(7)));
                     pet.nPoint.defg = Integer.parseInt(String.valueOf(dataArray.get(8)));
                     pet.nPoint.critg = Byte.parseByte(String.valueOf(dataArray.get(9)));
-                    int hp = Integer.parseInt(String.valueOf(dataArray.get(10)));
-                    int mp = Integer.parseInt(String.valueOf(dataArray.get(11)));
+                    long hp = Long.parseLong(String.valueOf(dataArray.get(10)));
+                    long mp = Long.parseLong(String.valueOf(dataArray.get(11)));
                     dataArray.clear();
                     // data body
                     dataArray = (JSONArray) jv.parse(String.valueOf(petData.get(2)));
